stable 9

+++ Emulator_test/sv_gps.h
@@ -20,11 +20,11 @@ namespace gps {
     geo::GEOPOSITION geoposition;
     
 //    quint32 course;
-    quint32 course_change_segment;
+    qreal course_change_segment;
     quint32 course_change_ratio;
     
 //    quint32 speed;
-    quint32 speed_change_segment;
+    qreal speed_change_segment;
     quint32 speed_change_ratio;
     
     quint32 roll_pitch_change_ratio;
